requirements.txt
asgiref==3.11.1
Django==5.2.11
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
dj-database-url
gunicorn
whitenoise
dj-database-url==2.2.0
gunicorn==23.0.0
